ReadBook Project

+++ SpMVC_Readbook/src/main/java/com/biz/book/mapper/BookDao.java
@@ -5,6 +5,6 @@ import com.biz.book.model.BookVO;
 /*
  * GenericDao를 extends 함으로써 기본 CRUD method를 별도로 정의하지 않아도 된다.
  */
-public interface BookDao extends GenericDao<BookVO, Integer>{
+public interface BookDao extends GenericDao<BookVO, Long>{
 	
 }
